renames logos
