+++ src/assets/images/footerlogoDarkWithURL.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import can_logo_png from './can_logo_url_dark.png';
+import can_logo_png from './covid-act-now-logo-url-dark.png';
 
 const LogoDarkWithURL = () => (
   <img src={can_logo_png} alt="Covid Act Now Logo"></img>
