+++ src/assets/images/logo.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import can_logo_png from './can_logo.png';
+import can_logo_png from './covid-act-now-logo.png';
 
 const Logo = ({ height }: { height?: number }) => (
   <img src={can_logo_png} height={height || 40} alt="Covid Act Now Logo"></img>
