+++ src/assets/images/logoDark.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import can_logo_dark_svg from './can_logo_dark.svg';
+import can_logo_dark_svg from './covid-act-now-logo-dark.svg';
 
 const Logo = ({ height }: { height?: number }) => (
   <img
